Change id (PK) database type from int to bigint

+++ backend/database/timesheet_database.sql
@@ -8,7 +8,7 @@ GRANT ALL PRIVILEGES ON DATABASE timesheet to timesheet;
 
 CREATE TABLE core_user
 (
-    id            INTEGER      NOT NULL PRIMARY KEY,
+    id            BIGINT      NOT NULL PRIMARY KEY,
     name          VARCHAR(250) NOT NULL,
     email         VARCHAR(250) NOT NULL,
     weekly_hours  FLOAT        NOT NULL,
@@ -21,7 +21,7 @@ CREATE TABLE core_user
 
 CREATE TABLE core_daily_time_sheet
 (
-    id      INTEGER NOT NULL PRIMARY KEY,
+    id      BIGINT NOT NULL PRIMARY KEY,
     date    DATE    NOT NULL,
     user_id INTEGER NOT NULL,
     CONSTRAINT fk_user
